requirements.txt
fastapi==0.112.1
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==0.19.0
celery==5.1.2
redis==3.5.3
requests==2.26.0
PyGithub==1.55
pydantic_core==2.20.1
click
cmd2